Update patch set 1
Patch Set 1:

(1 comment)

The class BufferedReader, in Java SE 7 and later, implements the interface java.lang.AutoCloseable. Because the BufferedReader instance is declared in a try-with-resource statement, it will be closed regardless of whether the try statement completes normally or abruptly (as a result of the method BufferedReader.readLine throwing an IOException).

https://docs.oracle.com/javase/tutorial/essential/exceptions/tryResourceClose.html

Patch-set: 1
Reviewer: Gerrit User 2571 <2571@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0